+++ CHANGELOG.md
@@ -9,6 +9,9 @@ was committed to the `master` branch. This is not necessarily the date that
 the code was deployed.
 
 ## [Unreleased]
+### Added
+- Added wifi link quality monitoring to Brave Hub in heartbeat.py (CU-fmy630).
+
 ### Security
 - Update Twilio references in order to get the latest version of axios (CU-j6yuzk).